Lax authentication in nginx
This method replaces the need to check the ip address of the caller.

Nginx strips the X-Consumer-Groups header, letting it through only if the caller satisfied the basic authentication. Only the api-gateway should know this secret.

This can be deployed without any blocker as api-gateway already passes the Authorization header. After that, Lax can remove checks for REMOTE_ADDR.